Reorganise applications components
