+++ apps/ui/components/application/ApplicationCard.tsx
