+++ apps/ui/components/application/ApplicationsGroup.tsx
