+++ apps/ui/components/application/funnel/ApplicationSectionSummary.tsx
