+++ apps/ui/components/application/funnel/ApplicationStepper.test.tsx
