+++ apps/ui/components/application/funnel/OrderedReservationUnitCard.tsx
@@ -273,6 +273,7 @@ export function OrderedReservationUnitCard({
   );
 }
 
+// TODO this is functioning as a base fragment for other cards, it should be renamed then
 export const ORDERED_RESERVATION_UNIT_CARD_FRAGMENT = gql`
   fragment OrderedReservationUnitCard on ReservationUnitNode {
     id
