+++ apps/ui/components/application/index.ts
@@ -1,25 +1,8 @@
 export { ApplicantInfoPreview } from "./ApplicantInfoPreview";
-export { ApplicantTypeSelector } from "./ApplicantTypeSelector";
-export { ApplicationFormTextInput } from "./ApplicationFormTextInput";
 export { ApplicationHead } from "./ApplicationHead";
-export { ApplicationFunnelWrapper } from "./ApplicationFunnelWrapper";
-export { ApplicationSection } from "./ApplicationSection";
-export { ApplicationStepper, PAGES_WITH_STEPPER } from "./ApplicationStepper";
 export { ApplicationTerms } from "./ApplicationTerms";
 export { ApprovedReservations } from "./ApprovedReservations";
-export { BillingAddress } from "./BillingAddress";
-export { CompanyForm, ContactPersonSection } from "./CompanyForm";
-export { IndividualForm } from "./IndividualForm";
 export { NotesWhenApplying } from "./NotesWhenApplying";
-export { OrderedReservationUnitCard } from "./OrderedReservationUnitCard";
-export { OrganisationForm } from "./OrganisationForm";
-export { Page1 } from "./Page1";
-export { Page2 } from "./Page2";
-export { type OptionTypes, ReservationUnitList } from "./ReservationUnitList";
-export { ReservationUnitModalContent } from "./ReservationUnitModalContent";
-export { TimePreview } from "./TimePreview";
-export { TimeSelector } from "./TimeSelector";
 export { ViewApplication } from "./ViewApplication";
 export { ApplicationSectionList } from "./ApplicationSectionList";
-export { ApplicationSectionSummary } from "./ApplicationSectionSummary";
-export { validateApplication } from "./form";
+export { ApplicationsGroup } from "./ApplicationsGroup";
